tested list token function

+++ backend/lib/forge-std/src/test/StdStorage.t.sol
@@ -13,7 +13,10 @@ contract StdStorageTest is Test {
     }
 
     function testStorageHidden() public {
-        assertEq(uint256(keccak256("my.random.var")), stdstore.target(address(test)).sig("hidden()").find());
+        assertEq(
+            uint256(keccak256("my.random.var")),
+            stdstore.target(address(test)).sig("hidden()").find()
+        );
     }
 
     function testStorageObvious() public {
@@ -47,7 +50,7 @@ contract StdStorageTest is Test {
             .with_key(address(this))
             .depth(1)
             .find();
-       assertEq(uint256(keccak256(abi.encode(address(this), 4))) + 1, slot);
+        assertEq(uint256(keccak256(abi.encode(address(this), 4))) + 1, slot);
     }
 
     function testStorageDeepMap() public {
@@ -57,7 +60,14 @@ contract StdStorageTest is Test {
             .with_key(address(this))
             .with_key(address(this))
             .find();
-        assertEq(uint256(keccak256(abi.encode(address(this), keccak256(abi.encode(address(this), uint(5)))))), slot);
+        assertEq(
+            uint256(
+                keccak256(
+                    abi.encode(address(this), keccak256(abi.encode(address(this), uint256(5))))
+                )
+            ),
+            slot
+        );
     }
 
     function testStorageCheckedWriteDeepMap() public {
@@ -78,7 +88,16 @@ contract StdStorageTest is Test {
             .with_key(address(this))
             .depth(0)
             .find();
-        assertEq(bytes32(uint256(keccak256(abi.encode(address(this), keccak256(abi.encode(address(this), uint(6)))))) + 0), bytes32(slot));
+        assertEq(
+            bytes32(
+                uint256(
+                    keccak256(
+                        abi.encode(address(this), keccak256(abi.encode(address(this), uint256(6))))
+                    )
+                ) + 0
+            ),
+            bytes32(slot)
+        );
     }
 
     function testStorageDeepMapStructB() public {
@@ -89,7 +108,16 @@ contract StdStorageTest is Test {
             .with_key(address(this))
             .depth(1)
             .find();
-        assertEq(bytes32(uint256(keccak256(abi.encode(address(this), keccak256(abi.encode(address(this), uint(6)))))) + 1), bytes32(slot));
+        assertEq(
+            bytes32(
+                uint256(
+                    keccak256(
+                        abi.encode(address(this), keccak256(abi.encode(address(this), uint256(6))))
+                    )
+                ) + 1
+            ),
+            bytes32(slot)
+        );
     }
 
     function testStorageCheckedWriteDeepMapStructA() public {
@@ -160,47 +188,73 @@ contract StdStorageTest is Test {
     }
 
     function testStorageCheckedWriteStructB() public {
-         stdstore.target(address(test)).sig(test.basic.selector).depth(1).checked_write(100);
+        stdstore.target(address(test)).sig(test.basic.selector).depth(1).checked_write(100);
         (uint256 a, uint256 b) = test.basic();
         assertEq(a, 1337);
         assertEq(b, 100);
     }
 
     function testStorageMapAddrFound() public {
-        uint256 slot = stdstore.target(address(test)).sig(test.map_addr.selector).with_key(address(this)).find();
-        assertEq(uint256(keccak256(abi.encode(address(this), uint(1)))), slot);
+        uint256 slot = stdstore
+            .target(address(test))
+            .sig(test.map_addr.selector)
+            .with_key(address(this))
+            .find();
+        assertEq(uint256(keccak256(abi.encode(address(this), uint256(1)))), slot);
     }
 
     function testStorageMapUintFound() public {
-        uint256 slot = stdstore.target(address(test)).sig(test.map_uint.selector).with_key(100).find();
-        assertEq(uint256(keccak256(abi.encode(100, uint(2)))), slot);
+        uint256 slot = stdstore
+            .target(address(test))
+            .sig(test.map_uint.selector)
+            .with_key(100)
+            .find();
+        assertEq(uint256(keccak256(abi.encode(100, uint256(2)))), slot);
     }
 
     function testStorageCheckedWriteMapUint() public {
-        stdstore.target(address(test)).sig(test.map_uint.selector).with_key(100).checked_write(100);
+        stdstore.target(address(test)).sig(test.map_uint.selector).with_key(100).checked_write(
+            100
+        );
         assertEq(100, test.map_uint(100));
     }
 
     function testStorageCheckedWriteMapAddr() public {
-        stdstore.target(address(test)).sig(test.map_addr.selector).with_key(address(this)).checked_write(100);
+        stdstore
+            .target(address(test))
+            .sig(test.map_addr.selector)
+            .with_key(address(this))
+            .checked_write(100);
         assertEq(100, test.map_addr(address(this)));
     }
 
     function testStorageCheckedWriteMapBool() public {
-        stdstore.target(address(test)).sig(test.map_bool.selector).with_key(address(this)).checked_write(true);
+        stdstore
+            .target(address(test))
+            .sig(test.map_bool.selector)
+            .with_key(address(this))
+            .checked_write(true);
         assertTrue(test.map_bool(address(this)));
     }
 
     function testFailStorageCheckedWriteMapPacked() public {
         // expect PackedSlot error but not external call so cant expectRevert
-        stdstore.target(address(test)).sig(test.read_struct_lower.selector).with_key(address(uint160(1337))).checked_write(100);
+        stdstore
+            .target(address(test))
+            .sig(test.read_struct_lower.selector)
+            .with_key(address(uint160(1337)))
+            .checked_write(100);
     }
 
     function testStorageCheckedWriteMapPackedSuccess() public {
         uint256 full = test.map_packed(address(1337));
         // keep upper 128, set lower 128 to 1337
         full = (full & (uint256((1 << 128) - 1) << 128)) | 1337;
-        stdstore.target(address(test)).sig(test.map_packed.selector).with_key(address(uint160(1337))).checked_write(full);
+        stdstore
+            .target(address(test))
+            .sig(test.map_packed.selector)
+            .with_key(address(uint160(1337)))
+            .checked_write(full);
         assertEq(1337, test.read_struct_lower(address(1337)));
     }
 
@@ -212,7 +266,7 @@ contract StdStorageTest is Test {
     function testFailStorageNativePack() public {
         stdstore.target(address(test)).sig(test.tA.selector).find();
         stdstore.target(address(test)).sig(test.tB.selector).find();
-        
+
         // these both would fail
         stdstore.target(address(test)).sig(test.tC.selector).find();
         stdstore.target(address(test)).sig(test.tD.selector).find();
@@ -234,7 +288,9 @@ contract StdStorageTest is Test {
     }
 
     function testStorageReadBool_Revert() public {
-        vm.expectRevert("stdStorage read_bool(StdStorage): Cannot decode. Make sure you are reading a bool.");
+        vm.expectRevert(
+            "stdStorage read_bool(StdStorage): Cannot decode. Make sure you are reading a bool."
+        );
         this.readNonBoolValue();
     }
 
@@ -271,10 +327,8 @@ contract StorageTest {
     uint248 public tA;
     bool public tB;
 
-
     bool public tC = false;
-    uint248 public tD = 1;    
-
+    uint248 public tD = 1;
 
     struct UnpackedStruct {
         uint256 a;
@@ -289,14 +343,11 @@ contract StorageTest {
     bool public tH = true;
 
     constructor() {
-        basic = UnpackedStruct({
-            a: 1337,
-            b: 1337
-        });
+        basic = UnpackedStruct({a: 1337, b: 1337});
 
-        uint256 two = (1<<128) | 1;
+        uint256 two = (1 << 128) | 1;
         map_packed[msg.sender] = two;
-        map_packed[address(bytes20(uint160(1337)))] = 1<<128;
+        map_packed[address(bytes20(uint160(1337)))] = 1 << 128;
     }
 
     function read_struct_upper(address who) public view returns (uint256) {
